Person model complete with single attribute (name) - all and find queries written
