+++ spec/test_stubs/person_statements.rb
@@ -1,4 +1,4 @@
-MEMBER_STATEMENTS = [
+PERSON_STATEMENTS = [
     RDF::Statement.new(RDF::URI.new('http://id.ukpds.org/member/1'), RDF::URI.new('http://schema.org/name'), 'Member1'),
     RDF::Statement.new(RDF::URI.new('http://id.ukpds.org/member/2'), RDF::URI.new('http://schema.org/name'), 'Member2'),
     RDF::Statement.new(RDF::URI.new('http://id.ukpds.org/member/3'), RDF::URI.new('http://schema.org/name'), 'Member3'),
@@ -6,7 +6,7 @@ MEMBER_STATEMENTS = [
     RDF::Statement.new(RDF::URI.new('http://id.ukpds.org/member/5'), RDF::URI.new('http://schema.org/name'), 'Member5'),
 ]
 
-MEMBER_ARRAY = [
+PERSON_ARRAY = [
     { id: '1',
       name: 'Member1'
     },
